Refactor: 피드삭제 함수 위치 변경 - 답변페이지에서 사용하므로 useAnswer훅으로 이동
