+++ src/pages/post/components/Questions.jsx
@@ -1,11 +1,18 @@
-import { useRouteLoaderData } from "react-router-dom";
-import { FeedCard, FeedCardList, FeedListHeader, FeedListWrapper } from "@components/FeedCard";
+import { useParams, useRouteLoaderData } from "react-router-dom";
+import {
+  FeedCard,
+  FeedCardList,
+  FeedDeleteButton,
+  FeedListHeader,
+  FeedListWrapper,
+} from "@components/FeedCard";
 import useLike from "./useLike";
 import useAnswer from "./useAnswer";
 
 export default function Questions({ count, data, mode = "view" }) {
   const userInfo = useRouteLoaderData("post");
-  const { create, update, remove, reject, isPending } = useAnswer();
+  const { id: subjectId } = useParams();
+  const { create, update, remove, reject, removeFeed, isPending } = useAnswer();
   const { mutate: reaction } = useLike();
 
   function handleCreate({ questionId, content }) {
@@ -37,25 +44,32 @@ export default function Questions({ count, data, mode = "view" }) {
     reaction({ questionId, type });
   }
 
+  function handleDeleteFeed() {
+    removeFeed({ subjectId });
+  }
+
   return (
-    <FeedListWrapper>
-      <FeedListHeader count={count} />
-      <FeedCardList data={data}>
-        {(question) => (
-          <FeedCard
-            key={question.id}
-            isPending={isPending}
-            question={question}
-            mode={mode}
-            feedOwner={userInfo}
-            onCreate={handleCreate}
-            onUpdate={handleUpdate}
-            onDelete={handleDelete}
-            onReject={handleReject}
-            onLike={handleLike}
-          />
-        )}
-      </FeedCardList>
-    </FeedListWrapper>
+    <>
+      {mode === "answer" && <FeedDeleteButton onClick={handleDeleteFeed} />}
+      <FeedListWrapper>
+        <FeedListHeader count={count} />
+        <FeedCardList data={data}>
+          {(question) => (
+            <FeedCard
+              key={question.id}
+              isPending={isPending}
+              question={question}
+              mode={mode}
+              feedOwner={userInfo}
+              onCreate={handleCreate}
+              onUpdate={handleUpdate}
+              onDelete={handleDelete}
+              onReject={handleReject}
+              onLike={handleLike}
+            />
+          )}
+        </FeedCardList>
+      </FeedListWrapper>
+    </>
   );
 }
